stock-breakout-agent/requirements.txt
anthropic>=0.40.0
pykrx>=1.0.47
playwright>=1.40.0
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2>=3.1.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pandas>=2.0.0
